feat: 페이지 컴포넌트 간 효율적인 상태 교류를 위한 context API 적용
- 모든 상태값에 context를 적용한 것이 아니라 페이지간 상태 교류를 위한 상태값들에만 context API 적용
