+++ src/App.jsx
@@ -1,6 +1,8 @@
 import { memo } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
 
+import { CardInfoProvider } from 'contexts/CardContext';
+
 import GlobalStyle from 'GlobalStyle';
 
 import { AddCard, AddCardComplete, CardList } from 'pages';
@@ -9,12 +11,14 @@ function App() {
   return (
     <>
       <GlobalStyle />
-      <Routes>
-        <Route path="/add-card" element={<AddCard />} />
-        <Route path="/add-card-complete" element={<AddCardComplete />} />
-        <Route path="/card-list" element={<CardList />} />
-        <Route path="*" element={<Navigate to="/card-list" replace />} />
-      </Routes>
+      <CardInfoProvider>
+        <Routes>
+          <Route path="/add-card" element={<AddCard />} />
+          <Route path="/add-card-complete" element={<AddCardComplete />} />
+          <Route path="/card-list" element={<CardList />} />
+          <Route path="*" element={<Navigate to="/card-list" replace />} />
+        </Routes>
+      </CardInfoProvider>
     </>
   );
 }
