+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Button from 'components/common/Button';
 import Card from 'components/common/Card';
+import CardForm from './CardForm';
 import CardKindButton from 'components/CardKindButton';
 import CardPickModal from 'components/CardPickModal';
 import CVCTooltip from 'components/CVCTooltip';
@@ -11,6 +12,7 @@ import ModalPortal from 'components/common/ModalPortal';
 export {
   Button,
   Card,
+  CardForm,
   CardKindButton,
   CardPickModal,
   CVCTooltip,
